Rework program logic into controller class.
