+++ gems.rb
@@ -16,7 +16,6 @@ end
 
 group :test do
 	gem "sus"
+	gem "covered"
 	gem "bake-test"
-	
-	gem "rspec"
 end
