+++ lib/bake/test/external.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+# Released under the MIT License.
+# Copyright, 2022, by Samuel Williams.
+
+require_relative 'external/version'
+require_relative 'external/controller'
